Time: 0 ms (100%), Space: 7.8 MB (68.85%) - LeetHub

+++ 1401-circle-and-rectangle-overlapping/1401-circle-and-rectangle-overlapping.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     bool checkOverlap(int radius, int xCenter, int yCenter, int x1, int y1, int x2, int y2) {
 
-        double dx = max(x1, min(xCenter, x2)) - xCenter;
-        double dy = max(y1, min(yCenter, y2)) - yCenter;
+        int dx = max(x1, min(xCenter, x2)) - xCenter;
+        int dy = max(y1, min(yCenter, y2)) - yCenter;
 
         return dx * dx + dy * dy <= radius * radius;
     }
